Add securityContext to controllers

+++ assign-api/src/main/java/com/robvangastel/assign/controllers/UserController.java
@@ -6,7 +6,7 @@ import com.robvangastel.assign.security.Secured;
 import com.robvangastel.assign.services.UserService;
 import io.swagger.annotations.Api;
 
-import javax.ejb.Stateless;
+import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 import javax.ws.rs.*;
 import javax.ws.rs.core.Context;
@@ -14,20 +14,13 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.SecurityContext;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.URI;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
 import java.util.List;
 
 /**
  * Created by Rob on 23-4-2017.
  */
 
-@Stateless
+@RequestScoped
 @Path("/users")
 @Api(tags = {"users"}, value = "/users", description = "Operations about users")
 @Produces({MediaType.APPLICATION_JSON})
@@ -37,7 +30,7 @@ public class UserController {
 	private UserService userService;
 
 	@Context
-	private SecurityContext context;
+	private SecurityContext securityContext;
 
 	@GET
 	public List<User> get() {
@@ -76,13 +69,12 @@ public class UserController {
 	}
 
 	@PUT
-	@Path("/{id}")
 	@Secured({Role.USER})
 	public Response update(@QueryParam("location") String location,
 					   @QueryParam("websiteURL") String websiteURL,
 					   @QueryParam("bio") String bio) throws Exception {
 		//TODO Update function
-		User user = userService.findByEmail(context.getUserPrincipal().getName());
+		User user = userService.findByEmail(securityContext.getUserPrincipal().getName());
 		if(user == null) {
 			throw new WebApplicationException(Response.Status.NOT_FOUND);
 		}
@@ -93,7 +85,12 @@ public class UserController {
 	@Path("/{id}")
 	@Secured({Role.ADMIN})
 	public Response delete(@PathParam("id") long id) throws Exception {
-		userService.delete(id);
-		return Response.noContent().build();
+		User user = userService.findByEmail(securityContext.getUserPrincipal().getName());
+		if(user.getRole().equals(Role.ADMIN.toString())) {
+			userService.delete(id);
+			return Response.noContent().build();
+		} else {
+			throw new WebApplicationException(Response.Status.UNAUTHORIZED);
+		}
 	}
 }
